Move to new AbstractConcurrentStreamBuilder
