+++ src/main/java/me/hugmanrique/riviere/AbstractConcurrentStreamBuilder.java
@@ -1,138 +1,247 @@
 package me.hugmanrique.riviere;
 
-import static java.util.Objects.requireNonNull;
-
-import java.util.Objects;
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicReferenceArray;
-import java.util.function.Supplier;
-import java.util.stream.IntStream;
+import java.lang.invoke.MethodHandles;
+import java.lang.invoke.VarHandle;
+import java.util.ArrayList;
+import java.util.Spliterator;
 import java.util.stream.Stream;
 
 /**
  * Concurrent {@code Stream} builder base implementation.
  *
- * @param <T> the type of the stream
- * @param <B> the type of the stream builders
+ * <p>This implementation employs a <em>non-blocking</em> algorithm based on
+ * one described in <a href="https://www.epfl.ch/labs/lamp/wp-content/uploads/2019/01/BrunoDIDOT.pdf">
+ * A Non-Blocking Concurrent Queue Algorithm</a> by B. Didot, with support
+ * for increasing node capacities and primitive types.
+ *
+ * <p>This significantly outperforms implementations based on
+ * serialized access to a builder returned by {@link Stream#builder()}
+ * (and primitive equivalents) and a synchronized {@link ArrayList}
+ * used as a temporary buffer (which additionally introduces
+ * copying overhead).
+ *
+ * <p>The maximum capacity of an {@link AbstractConcurrentStreamBuilder}
+ * is {@code 2^(32 - log2(k)) - 1} where {@code k} is the initial capacity.
+ *
+ * @param <A> the array type for the stream element type
+ * @param <S> the supplier type for the stream element type
  */
-abstract class AbstractConcurrentStreamBuilder<T, B> {
+abstract class AbstractConcurrentStreamBuilder<A, S> {
 
-    // This base implementation lazily allocates a set number of
-    // buckets, i.e. thread-unsafe stream builders. Each caller is
-    // assigned a particular bucket based on the current thread's
-    // hashcode. The retrieval (and initialization) of the bucket
-    // is performed atomically. However, to ensure memory visibility
-    // the callers must acquire a lock on the bucket itself.
-    // This lock is non-contested, making the acquisition overhead
-    // negligible on modern JVMs.
+    // This is a modification of the enqueuing algorithm based on
+    // a singly linked unrolled list proposed by B. Didot,
+    // with support for variable node capacities.
     //
-    // The buckets are not dereferenced after building the final stream.
-    // This is based on the assumption that builder objects are short-lived
-    // and are garbage collected after calling build().
+    // The fundamental invariants are:
+    // - There is exactly one (last) Node with a null next reference,
+    //   which is CASed when enqueuing. It can always be reached from
+    //   the head.
+    // - The elements contained in the builder are the initialized
+    //   elements in the items array of each Node.
+    // - All predecessor items of an initialized element are initialized.
+    //   This includes all the items in a Node that can reach
+    //   the Node of that element.
+    // - Nodes cannot be dequeued. The list can only be traversed if
+    //   the builder is in the built state.
     //
-    // The level of expected concurrency can be customized via
-    // the bucketCount parameter. It defaults to the number of
-    // available processors.
-
-    private static final int MAXIMUM_BUCKET_COUNT = Runtime.getRuntime().availableProcessors();
-
-    private static final int DEFAULT_BUCKET_COUNT = 16; // ConcurrentHashMap.DEFAULT_CAPACITY
-
-    private final Supplier<? extends B> bucketSupplier;
-
-    private final AtomicReferenceArray<B> buckets;
-    private final int bucketCount;
-
-    private final AtomicBoolean isBuilt = new AtomicBoolean(false);
+    // The tail is permitted to lag. Both head and tail may point to
+    // an empty Node. Since no dequeuing is performed, tail cannot lag
+    // behind head, so head need not be a dummy node as in
+    // the original implementation.
 
     /**
-     * Constructs a concurrent stream builder with {@link #DEFAULT_BUCKET_COUNT} buckets.
-     * The supplier should be side-effect-free, since it may be called when
-     * attempted updates fail due to contention among threads.
-     *
-     * @param bucketSupplier the side-effect-free stream builder supplier
+     * The default initial capacity of the first {@link Node} in
+     * a {@link AbstractConcurrentStreamBuilder}.
      */
-    protected AbstractConcurrentStreamBuilder(final Supplier<? extends B> bucketSupplier) {
-        this(bucketSupplier, DEFAULT_BUCKET_COUNT);
-    }
+    private static final int DEFAULT_INITIAL_CAPACITY = 16;
 
-    /**
-     * Constructs a concurrent stream builder with the given number of maximum buckets.
-     * The supplier should be side-effect-free, since it may be called when
-     * attempted updates fail due to contention among threads.
-     *
-     * @param bucketSupplier the side-effect-free stream builder supplier
-     * @param bucketCount the maximum number of underlying stream builders
-     * @throws IllegalArgumentException if {@code bucketCount} is non-positive
-     */
-    protected AbstractConcurrentStreamBuilder(final Supplier<? extends B> bucketSupplier,
-                                              int bucketCount) {
-        if (bucketCount <= 0) {
-            throw new IllegalArgumentException("Bucket count is non-positive, got " + bucketCount);
+    abstract static class Node<A, S> {
+        private volatile Node<A, S> next;
+        protected final A items;
+        private final int capacity;
+
+        /**
+         * A value greater than or equal to the number of initialized items.
+         *
+         * <p>The original implementation supports weak updates to
+         * an {@code addHint} counter used to speed up the last non-null
+         * element search. However, Node must support primitive arrays
+         * with no {@code null} equivalent (a magic constant is not
+         * a valid solution). Instead, {@code count} is atomically
+         * incremented every time an item is added to the Node.
+         * This value can technically overflow, but this isn't
+         * a problem in practice.
+         *
+         * <p>If {@code count >= items.length}, the Node is full.
+         */
+        private volatile int count;
+
+        /**
+         * Constructs an empty node that can hold {@code capacity} items.
+         *
+         * @param capacity the node capacity
+         */
+        Node(final int capacity) {
+            if (capacity <= 0)
+                throw new IllegalArgumentException("Got non-positive capacity " + capacity);
+            this.capacity = capacity;
+            this.items = newArray(capacity);
         }
-        this.bucketSupplier = requireNonNull(bucketSupplier);
-        this.bucketCount = Math.min(bucketCount, MAXIMUM_BUCKET_COUNT);
-        this.buckets = new AtomicReferenceArray<>(this.bucketCount);
+
+        /**
+         * Constructs a node that can hold {@code capacity} items
+         * whose {@code expectedCount} first items are to be
+         * initialized by the parent class constructor.
+         *
+         * @param capacity the node capacity
+         * @param expectedCount the number of initialized items
+         */
+        protected Node(final int capacity, final int expectedCount) {
+            this(capacity);
+            // Relaxed write, we piggyback on publication via CAS
+            COUNT.set(this, expectedCount);
+        }
+
+        protected abstract A newArray(int length);
+
+        protected abstract void setVolatile(final int index, final S supplier);
     }
+
+    /**
+     * Indicates the builder is in built state, at which point no
+     * new elements may be enqueued and the list can be traversed.
+     */
+    private volatile boolean built;
+
+    /**
+     * The first node in the list (may contain items).
+     */
+    private final Node<A, S> head;
+
+    /**
+     * A node from which the last node on the list can be reached.
+     */
+    private volatile Node<A, S> tail;
+
+    protected AbstractConcurrentStreamBuilder() {
+        this(DEFAULT_INITIAL_CAPACITY);
+    }
+
+    protected AbstractConcurrentStreamBuilder(final int initialCapacity) {
+        head = tail = createEmptyNode(initialCapacity);
+    }
+
+    protected abstract Node<A, S> createEmptyNode(final int capacity);
+
+    protected abstract Node<A, S> createNextNode(final int capacity, final S valueSupplier);
 
     private void throwBuilt() {
         throw new IllegalStateException("Builder is in built state");
     }
 
-    /**
-     * Returns the builder bucket associated to the current thread.
-     *
-     * @return the current stream builder
-     */
-    final B get() {
-        if (isBuilt.get()) {
-            throwBuilt();
-        }
-
-        int index = Thread.currentThread().hashCode() % bucketCount;
-        return buckets.updateAndGet(index, previous ->
-                previous != null ? previous : bucketSupplier.get());
+    private boolean isBuilt() {
+        // Don't impose any ordering constraints w.r.t. other variables
+        return (boolean) BUILT.getOpaque(this);
     }
 
-    /**
-     * Builds the stream, transitioning this builder to the built state.
-     *
-     * @return the built stream
-     * @throws IllegalStateException if the builder has already transitioned to
-     *         the built state
-     */
-    public final T build() {
-        if (!isBuilt.compareAndSet(false, true)) {
+    protected void checkAndSetBuilt() {
+        if (!BUILT.compareAndSet(this, false, true))
             throwBuilt();
-        }
+    }
 
-        Stream<T> streams = IntStream.range(0, bucketCount)
-                .mapToObj(buckets::get)
-                .filter(Objects::nonNull)
-                .map(builder -> {
-                    //noinspection SynchronizationOnLocalVariableOrMethodParameter
-                    synchronized (builder) {
-                        return build0(builder);
+    protected void enqueue(final S valueSupplier) {
+        Node<A, S> nextNode = null;
+        while (!isBuilt()) {
+            Node<A, S> curTail = tail;
+            Node<A, S> tailNext = curTail.next;
+            if (tailNext == null) {
+                // curTail is last node
+                int index = (int) COUNT.getAndAdd(curTail, 1);
+                if (index < curTail.capacity) {
+                    curTail.setVolatile(index, valueSupplier);
+                    return;
+                } else {
+                    // The node is full (was already full or lost CAS race).
+                    // Create next node (if necessary) and try to append.
+                    int nextCap = curTail.capacity << 1; // overflow checked by Node constructor
+                    if (nextNode == null || nextNode.capacity != nextCap) {
+                        nextNode = createNextNode(nextCap, valueSupplier);
                     }
-                });
 
-        return flatMap(streams);
+                    if (NEXT.compareAndSet(curTail, null, nextNode)) {
+                        // If this CAS fails, another caller will advance tail
+                        TAIL.compareAndSet(this, curTail, nextNode);
+                        return;
+                    }
+                }
+            } else {
+                // Help advance tail
+                TAIL.compareAndSet(this, curTail, tailNext);
+            }
+        }
+        throwBuilt();
     }
 
-    /**
-     * Builds a stream from the given builder.
-     *
-     * @param builder the builder
-     * @return the built stream
-     */
-    abstract T build0(B builder);
+    protected abstract class AbstractSpliterator<T extends Spliterator<?>> {
+        protected Node<A, S> node;
+        protected int index; // current index in node
 
-    /**
-     * Returns a stream consisting of the elements of all streams
-     * contained in the given stream.
-     *
-     * @param streams the streams containing the elements
-     * @return the new stream
-     */
-    abstract T flatMap(Stream<T> streams);
+        private long est = -1; // size estimate, -1 until needed
+
+        public AbstractSpliterator() {
+            if (!isBuilt())
+                throw new AssertionError(
+                        "Spliterator constructed while builder is not in built state");
+            this.node = head;
+        }
+
+        public T trySplit() {
+            return null;
+        }
+
+        public long estimateSize() {
+            if (est < 0) {
+                // Traverse list to sum up counts
+                // TODO Nodes have increasing power of 2 capacities, we could
+                // compute the size of all nodes except the last with a shift.
+                Node<A, S> current = head;
+                do {
+                    est += current.count;
+                } while ((current = current.next) != null);
+            }
+            return est;
+        }
+
+        protected boolean advance() {
+            if (node == null) return false;
+            if (index >= node.capacity) {
+                node = node.next;
+                index = 0;
+                if (node == null) return false;
+            }
+            return index < node.count;
+        }
+
+        public int characteristics() {
+            return Spliterator.ORDERED | Spliterator.SIZED | Spliterator.SUBSIZED;
+        }
+    }
+
+    // VarHandle mechanics
+    private static final VarHandle TAIL;
+    private static final VarHandle BUILT;
+    private static final VarHandle NEXT;
+    private static final VarHandle COUNT;
+    static {
+        try {
+            MethodHandles.Lookup l = MethodHandles.lookup();
+            TAIL = l.findVarHandle(AbstractConcurrentStreamBuilder.class, "tail", Node.class);
+            BUILT = l.findVarHandle(
+                    AbstractConcurrentStreamBuilder.class, "built", boolean.class);
+            NEXT = l.findVarHandle(Node.class, "next", Node.class);
+            COUNT = l.findVarHandle(Node.class, "count", int.class);
+        } catch (final ReflectiveOperationException e) {
+            throw new ExceptionInInitializerError(e);
+        }
+    }
 }
