+++ src/main/java/me/hugmanrique/riviere/ConcurrentIntStreamBuilder.java
@@ -1,47 +1,92 @@
 package me.hugmanrique.riviere;
 
+import java.lang.invoke.MethodHandles;
+import java.lang.invoke.VarHandle;
+import java.util.Objects;
+import java.util.Spliterator;
+import java.util.function.IntConsumer;
+import java.util.function.IntSupplier;
 import java.util.stream.IntStream;
-import java.util.stream.Stream;
+import java.util.stream.StreamSupport;
 
 /**
  * A {@link IntStream.Builder} supporting full concurrency of additions.
  */
-@SuppressWarnings("SynchronizationOnLocalVariableOrMethodParameter")
-public final class ConcurrentIntStreamBuilder extends AbstractConcurrentStreamBuilder<IntStream, IntStream.Builder>
-        implements IntStream.Builder {
+public final class ConcurrentIntStreamBuilder
+        extends AbstractConcurrentStreamBuilder<int[], IntSupplier> implements IntStream.Builder {
+
+    private static final VarHandle ITEM = MethodHandles.arrayElementVarHandle(int[].class);
+
+    private static final class IntNode extends Node<int[], IntSupplier> {
+
+        private IntNode(final int capacity) {
+            super(capacity);
+        }
+
+        private IntNode(final int capacity, final int firstItem) {
+            super(capacity, 1);
+            // Piggyback on publication via CAS
+            ITEM.set(this.items, 0, firstItem);
+        }
+
+        @Override
+        protected int[] newArray(final int length) {
+            return new int[length];
+        }
+
+        @Override
+        protected void setVolatile(final int index, final IntSupplier supplier) {
+            ITEM.setVolatile(this.items, index, supplier.getAsInt());
+        }
+    }
 
     /**
      * Constructs a concurrent {@link IntStream} builder.
      */
-    public ConcurrentIntStreamBuilder() {
-        super(IntStream::builder);
-    }
+    public ConcurrentIntStreamBuilder() {}
 
     /**
-     * Constructs a concurrent {@link IntStream} builder with the given number
-     * of maximum buckets.
+     * Constructs a concurrent {@link IntStream} builder with
+     * the given initial node capacity.
      *
-     * @param bucketCount the maximum number of underlying stream builders
+     * @param initialCapacity the capacity of the head node
      */
-    public ConcurrentIntStreamBuilder(final int bucketCount) {
-        super(IntStream::builder, bucketCount);
+    public ConcurrentIntStreamBuilder(final int initialCapacity) {
+        super(initialCapacity);
+    }
+
+    @Override
+    protected IntNode createEmptyNode(final int capacity) {
+        return new IntNode(capacity);
+    }
+
+    @Override
+    protected IntNode createNextNode(final int capacity, final IntSupplier valueSupplier) {
+        return new IntNode(capacity, valueSupplier.getAsInt());
     }
 
     @Override
     public void accept(final int value) {
-        var builder = get();
-        synchronized (builder) {
-            builder.accept(value);
+        enqueue(() -> value);
+    }
+
+    @Override
+    public IntStream build() {
+        checkAndSetBuilt();
+        return StreamSupport.intStream(new BuilderSpliterator(), false);
+    }
+
+    private final class BuilderSpliterator extends AbstractSpliterator<Spliterator.OfInt>
+            implements Spliterator.OfInt {
+
+        @Override
+        public boolean tryAdvance(final IntConsumer action) {
+            Objects.requireNonNull(action);
+            boolean canAdvance = advance();
+            if (canAdvance) {
+                action.accept(node.items[index]);
+            }
+            return canAdvance;
         }
-    }
-
-    @Override
-    IntStream build0(final IntStream.Builder builder) {
-        return builder.build();
-    }
-
-    @Override
-    IntStream flatMap(final Stream<IntStream> streams) {
-        return streams.flatMapToInt(s -> s);
     }
 }
